feature/add nfs vlan (#1050)
* Add node-labeler app for labeling nodes by primary interface name

This applies the massopen.cloud/primary-interface label nodes so that
we can use it for selecting an appropriate network configuration.

Part of operate-first/SRE#386

* Add nmstate configuration for nfs vlan

This adds the nfs vlan to all worker nodes. There are multiple
configuration because our worker nodes don't have consistent interface
names.

This requires the massopen.cloud/primary-interface label to select the
appropriate nmstate configuration for a node.

Part of operate-first/SRE#386