Add scripts/fake-ksops.sh to ease testing w/o ksops
We have a number of new co-ops working on this project. Testing out
changes by running `kustomize build` is complicated by our use of ksops.
Rather than distributing decryption keys to all and sundry, we can ease
the testing process by providing a wrapper that sets up a fake ksops
plugin.

The script is used like this:

    ./scripts/fake-ksops.sh kustomize build --enable-alpha-plugins ...

That permits one to test changes that do not involve modifications to
encrypted data.