Installing volsync using kustomize (#1609)
* Add volsync template generator

This adds a kustomization file that will explode the volsync helm
chart into a set of manifests.

* Add volsync cluster-scoped resources

Add cluster-scoped resources for volsync (custom resource definitions,
cluster roles and bindings, etc).

* Add volsync namespaced resources

* Add smaug overlay for volsync

This does nothing but include the base manifests, but it makes the
volsync application match the layout of our other argocd apps.

* Add volsync namespace

The helm template doesn't create the target volsync-system namespace,
so we add it explicitly.

* Add missing volsync scc

The helm chart conditionally generates an SCC if it detects it is
being installed in OpenShift, but this mechanism doesn't work when
we're exploding a template via 'helm template' (or equivalent using
Kustomize).

* Add volsync application to argocd

Replace the previous helm-chart based application with one that
follows our standard practice of deploying via kustomize manifests.

* Add volsync bundle

Add a resource bundle for all the volsync cluster scoped resources.

* Add volsync to smaug overlay

Include volsync cluster-scoped resources in smaug's overlay (by
including the volsync resource bundle).