Move idp auth resources to separate namespace.
This update revises the structure for acm placements a bit

IDP authrealm and placement moved to separate namespace,
as the management of these resources is a separate concern.

ArgoCD and IDP placement setup is affected. ArgoCD integration
is simpliefied a little bit. Users would now only need to
create a ManagedClusterSetBinding in one location and it should
be deployed in both idp/argocd namespaces. ArgoCD placement
will auto pick up any ManagedCluster bounded to its namespace,
while IDP will be "enabled" but the Managed cluster will still
need a specific label set to configure the associated cluster
with the configured Authrealm (in this case opf github org).

Docs were updated to reflect this change.