Add oc pipelines as a bundle to os-cl2.
Add oc pipelines to osc-cl2. This commit also bundles the manifests that
currently exist in the base dir and switches to bundle reference in the
smaug cluster as well. No change to live deployment to the smaug cluster
is introduced here. The patch for the subscription was removed since the
channel remains the same for both clusters, it just adds an unecessary
step otherwise.