Add auto sync of notebooks to hitchhiker repo.
This will enable automated syncs via gh actions whenever a notebook is
updated. This is because we can only configure one image build per repo,
and it does not make sense to dedicate the apps repo for a notebook
image build. So instead we'll have a separate repo where we initiate our
builds from, and this action will ensure that while notebooks can be
committed to this repo, they will continue to initiate new image builds
by auto syncing with the build repository.