Continued efforts to sync trino-acl-dsl.yaml and rules.json
Fix end-of-line problem, plus work-around for the fact that it's not possible to fix the yaml file without there being a corresponding update of the rules.json file.  In other words, if you accidentally do the right thing the wrong way, the only way forward is to add something arbitrary that can be removed later.