Add pachyderm operator and deployment to osc-cl2.
This PR creates a new bundle in base directory comprising of the
pachyderm operator subscription and the namespace to which pachyderm
will be deployed. The bundle is included in the osc-cl2 cluster
kustomization. There is also included a pachyderm custom resource
deployment that will instruct the operator to deploy pachyderm into
the namespace. An argocd application that will deploy the newly
introduced pachyderm custom resources to the osc-cl2 is also included.