Update smaug 1.1.2 (#1873)
* Rename odh component folders to ensure consistency.

This change will ensure odh component folder names for smaug are
consistent with that of the other clusters. This will make it easier
when adding generic odh related docs and paths need to be specified.

* Update smaug odh kfdefs to 1.1.2.

This will switch odh smaug kfdefs to deploy from the smaug-v1.1.12
odh-manifests branch. Some changes to superset and how its oauth secret
and superset-config.py file are managed are made. Superset config file
is now managed from the apps repo via argocd as it often requires
frequent changes, and it makes little sense to have odh manage this for
us. The oauth secret is also pulled into the same secret where superset
admin credentials are specified, this set up makes it more consistent
with how we manage files in the osc clusters.