Set query.max-length=10M (overriding 1M default) (#1883)
The default query.max-length=1000000.  For datasets such as SEC_DERA, the 1M limit allows for approx 1500 rows per insertion.  When there are millions of rows, this not only creates thousands of snapshots, but each snapshot file also references the chain of snapshots, growing these files from a few K to several MB per snapshot.

By increasing the max-length parameter to 10M, we not only cut down the number of objects in the metastore, but we also decrease the number of snapshots each of these objects needs to reference, reducing storage requirements by a factor of 100 (10*10).