Route nooba alerts to slack in smaug.
Alertmanager config requires you to put the slack api url within the
alertmanager secret. This is unfortunate because we would like to keep
most of the alertmanager secret visible to the public as all our routing
and inhibit rules go there.

To get around this, we build the alert manager secret using ESO by
pulling from the alert-manager-template configmap. ESO will patch in the
slack api url from vault upon deployment and create the alertmanager-main
secret in openshift-monitoring.

To better accomodate this change the commit restructures configmaps &
secrets within the smaug overlay so that they are confined to their
respective sub folders.