Add eso to all remainings clusters.
This change will deploy ESO to all clusters (save for smaug which
already has it). And set up an initial secret store for the namespace
openshift-monitoring, as a means to confirm it works and also as a set
up for a future addition of slack alerting support (the slack api url
will need to be fetched by eso from vault for all clusters).