This commit restructures the cert-manager folder.
Base directory and all forms of patching is removed. This is because
there was an extensive degree of patching taking place making adding
additional changes/clusters needlessly complicated. There is also no
clear pattern/commonality between all clusters. Some clusters in the
same envrionment/region are using either route53 or google dns and
their sharing of manifests doesn't make much sense. As a result there
is now some repetition, but it is much easier to understand the layout,
making the trade-off acceptable.