Dex secure (#2052)
* Leverage ESO for dex and secure routes.

This change will switch to using external secrets to fetch client
secrets for our dex installlation for all clusters. To accomodate this
change, the dex folder structure is also overhauled and made consistent
across the various base/overlay sub directories. Configurations can now
be updated in their respective files, and are pulled in via the
configmapgenerator.

The new structure also leverages ESO to pull secrets for dex. To
accomodate this a secretstore has been added, SOPS secrets have bene
removed, and instead external secrets have been added as a replacement.

* Make dex routes secure.

* Add dex auth service for trino to osc-cl1.

This manifests are already deployed to the trino namespace but were not
added to git. The only additional change being made here to live is that
the dex routes are switched to https.