Add backup job for vault.
This PR adds a backup job using openshift pipelines. The backup job
makes use of a taskrun that will exec onto hashicorp vault on smaug
cluster and create a snapshot, which is then copied to pvc and s3
storage.

Documentation on how to use this snapshot to restore vault are also
included, as well as details on the set up of this task. There is also a
pod manifest provided that one can use to access the pvc and retrieve
backed up snapshots.