Add secretstores to all namespaces using eso.
This change accomplishes 2 things:

1) Move all existing secret stores to cluster overlays, and makes their
   usage pattern consistent across clusters.
2) Add secretstores to all namespaces that currently require encrypted
   secrets to be stored in the apps repo.

Majority of the changes consist of (2). The only real affect to the live
environment is renaming some of the secret stores so that they are
consistent with the others (this requires updating the external secrets
themselves). The remainder of the secret stores only add secret stores,
and thus have no real affect to the existing environment. Once secret
stores are in place, we will be able to replace all sops encrypted
secrets with external secrets for all these namespaces.