Add aiops tools workshop manifests.
The manifests introduce various different services into the
aiops-tools-workshop namespace. This includes, odh, trino, superset,
jupyterhub, cloubeaver, dex, kfp-tekton, various configuration files for
these services in the form of configmaps and externalsecrets.

A readme is included to offer further insight into the dependencies
required. The smaug deployment already has some dependencies included in
the base directory for the workshop, so they are excluded from the smaug
overlay. However, when looking to deploy this workshop to another
cluster, ensure these dependencies are either 1) already included, or 2)
included within this overlay. Otherwise the workshop will not work.