Add kepler to osc-cl1 cluster.
This PR reorganizes the machineconfigs and pulls them out of the kepler
bundle. As each clusters node status is different, and the order of
application can vary, we cannot presume that the same naming scheme will
apply to all clusters. Furthermore, we cannot assume that we want kepler
deployed on the same set of nodes. For instance in the case of osc-cl1
we only want to deploy to non GPU nodes with label "osc-platform" while
for morty we deployed to all workers and masters. As such we need finer
cluster level specifications in their overlays. These must exist in the
overlays, because we cannot exclude resources via kustomize.

The PR also adds the relevant argocd app to deploy the namespace scoped
resources for kepler to osc-cl1.