Pin custom elyra image until 0.2.x bug is fixed (#2617)
The most recent versions of the image, 0.2.0 and 0.2.1, have a bug
that is causing the notebook to fail to load. An issue has been
opened with the upstream project, but until it is fixed and new image
is released, we need to pin to last working version.
Fixes operate-first/support#1137

Signed-off-by: Eric Ball <eball@linuxfoundation.org>

Signed-off-by: Eric Ball <eball@linuxfoundation.org>